switch to double + written distribution func

+++ SaluteStocksAPI/Models/Distribution/DistributionValue.cs
@@ -2,10 +2,10 @@ namespace SaluteStocksAPI.Models.Distribution;
 
 public struct DistributionValue
 {
-    public decimal Position;
+    public double Position;
     public int Value;
 
-    public DistributionValue(KeyValuePair<decimal, int> kv)
+    public DistributionValue(KeyValuePair<double, int> kv)
     {
         Position = kv.Key;
         Value = kv.Value;
